refactor: modify layout more reusable
